pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "market-analyst"
version = "0.1.0"
description = "AI-powered market analysis — news-driven causal reasoning with stock picks"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "rich>=13.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "yfinance>=0.2.36",
    "scipy>=1.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.scripts]
server = "app.api.server:main"
cli = "app.cli.main:main"
